lt-42 / Рефакторинг
