+++ packages/client/src/pages/Leaderboard/ui/LeaderboardItem/index.tsx
@@ -1,23 +1,26 @@
 import type { FC } from 'react';
 import { TLeaderboardItemProps } from './types';
+
 import css from './styes.module.scss';
 
 export const LeaderboardItem: FC<TLeaderboardItemProps> = props => {
   const { avatarUrl, nickName, score, dynamics, place } = props;
   return (
-    <div className={css.container}>
-      <div className={css.avatarWrapper}>
+    <div className={css['leader-card']}>
+      <div className={css['leader-card__img-wrapper']}>
         <img
-          className={css.avatarImg}
+          className={css['leader-card__img']}
           src={avatarUrl}
           alt={`${nickName}_avatar`}
         />
       </div>
 
-      <span className={css.nickName}>{nickName}</span>
-      <span className={css.score}>{score}</span>
-      <span className={`${css.dynamics} dynamics-${dynamics}`} />
-      <span className={css.place}>{place}</span>
+      <span className={css['leader-card__name']}>{nickName}</span>
+      <span className={css['leader-card__score']}>{score}</span>
+      <span
+        className={`${css['leader-card__dynamics']} dynamics-${dynamics}`}
+      />
+      <span className={css['leader-card__place']}>{place}</span>
     </div>
   );
 };
